update tests and merge all color ramp tests into single file
